pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-search-mcp"
version = "0.1.3"
authors = [
  { name = "P.S Zhang", email = "pengsongzhang96@gmail.com" },
]
description = "A MCP server for searching and downloading academic papers from multiple sources."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "feedparser",
    "fastmcp",
    "mcp[cli]>=1.6.0",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0", # Better HTML parser for BeautifulSoup
    "httpx[socks]>=0.28.1",
    "docling>=2.66.0",
    "sentence-transformers>=5.2.0",
    "numpy>=2.2.6",
    "faiss-cpu>=1.13.2",
]

[tool.hatch.build.targets.wheel]
packages = ["paper_search_mcp"]
